Created some animations for starting and a score system

+++ ColorMatch_Game_Test/Assets/Scripts/FloatData.cs
@@ -1,10 +1,11 @@
+using System;
 using UnityEngine; 
 [CreateAssetMenu] 
-public class IntData : ScriptableObject 
+public class FloatData : ScriptableObject 
 { 
     public float value;
 
-    public void SetValue(int num)
+    public void SetValue(float num)
     {
         value = num;
     }
